+++ src/main/java/executor/service/service/stepexecution/StepExecutionSleep.java
@@ -1,4 +1,4 @@
-package executor.service.service;
+package executor.service.service.stepexecution;
 
 import executor.service.model.Step;
 import org.openqa.selenium.WebDriver;
