+++ src/main/java/executor/service/service/stepexecution/impl/ScenarioExecutorImpl.java
@@ -1,10 +1,7 @@
-package executor.service.service.impl;
+package executor.service.service.stepexecution.impl;
 
 import executor.service.model.Scenario;
 import executor.service.service.ScenarioExecutor;
-import executor.service.service.StepExecutionClickCss;
-import executor.service.service.StepExecutionClickXpath;
-import executor.service.service.StepExecutionSleep;
 import org.openqa.selenium.WebDriver;
 
 public class ScenarioExecutorImpl implements ScenarioExecutor {
