+++ src/main/java/executor/service/service/stepexecution/impl/StepExecutionClickCssImpl.java
@@ -1,8 +1,8 @@
-package executor.service.service.impl;
+package executor.service.service.stepexecution.impl;
 
 import executor.service.model.Step;
 import executor.service.model.StepTypes;
-import executor.service.service.StepExecutionClickCss;
+import executor.service.service.stepexecution.StepExecutionClickCss;
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
