+++ src/main/java/executor/service/service/stepexecution/impl/StepExecutionFactory.java
@@ -1,8 +1,8 @@
-package executor.service.service.impl;
+package executor.service.service.stepexecution.impl;
 
 import executor.service.exceptions.StepExecutionException;
-import executor.service.service.StepExecution;
-import executor.service.service.StepExecutionFabric;
+import executor.service.service.stepexecution.StepExecution;
+import executor.service.service.stepexecution.StepExecutionFabric;
 import java.util.List;
 
 public class StepExecutionFactory implements StepExecutionFabric {
